Initial commit of OS X console sample
This is the initial commit of the sample code that I have written to
allow the control of the MiP robot over Bluetooth Low Energy (BLE)
using the Core Bluetooth framework on OS X.

Two threads are used in this sample.  The main thread is left executing
a NSApplication run loop.  A worker thread is started to run the
code which wants to interact with the robot to do things like
commanding it to drive around.  The code running in this thread
starts at the robotMain() function found in main.c  When it needs
to utilize the Core Bluetooth stack, it communicates its requests to
the main thread via mipTransport*() functions.

There are a few limitations to the current code:
* The transport mechanism, OS X Core Bluetooth in this case, is exposed
  directly to the developer.  In the future this will be abstracted
  behind a transport agnostic interface.  But that will only happen if
  I continue to experiment with the MiP robot after playing around with
  this current API to see what is and isn't possible to accomplish with
  this little bot.
* There is currently no code to attempt reconnecting to the MiP if it
  should go out of range or otherwise lose its BLE connectivity for a
  brief period of time.
* It is relatively untested at this time.  I will continue to test and
  explore the abilities of the MiP and this code in the near futre.
* There is limited documentation.  I did add some comments to the
  mip.h header file.  I will flush this documentation out further in
  the future if I decide to continue with these MiP hacks.