Add first high level MiP functions
I added some high level MiP functions to handle the operations used by
the wall following sample so that it no longer needs to use the
mipRaw*() functions.  The added functions include:
* mipSetGestureRadarMode()
* mipGetGestureRadarMode()
* mipSetChestLED()
* mipContinuousDrive()
* mipTurnLeft()
* mipTurnRight()
* mipStop()
* mipGetLatestRadarNotification()

I also added documentation and examples for these functions.  This
commit is really the beginning of the MiP C API.