Add MiP status functions
Following functions added:
* mipGetStatus()
* mipGetLatestStatusNotification()

I also renamed mipSetPosition() to mipFallDown() to make it more
descriptive of what it actually does.  The original MiPPosition enum
was renamed to MiPFallDirection so that the name MiPPosition could be
used for the robot position returned in the MiP status report.  Some
of the examples and documentation had to be updated due to these
changes.