Move MiP C API to own repo
I have moved the MiP C API library to its own repo
 git@github.com:adamgreen/MiP-Capi.git

This commit deletes the library code from this repo and instead uses
the new repo as a git submodule.  The updated
OSX_ConsoleSample/makefile is a slightly modified version of the one I
created in the MiP-Capi standalone project.